+++ packages/g-plugin-device-renderer/src/interfaces.ts
@@ -17,3 +17,7 @@ export interface RendererParameters {
   toneMapping: ToneMapping;
   toneMappingExposure: number;
 }
+
+export interface DeviceRendererPluginOptions {
+  enableFXAA: boolean;
+}