+++ packages/g-webgpu/src/index.ts
@@ -1,5 +1,4 @@
-import { ClipSpaceNearZ, RendererConfig } from '@antv/g-lite';
-import { AbstractRenderer } from '@antv/g-lite';
+import { AbstractRenderer, ClipSpaceNearZ, RendererConfig } from '@antv/g-lite';
 import * as DeviceRenderer from '@antv/g-plugin-device-renderer';
 import * as DomInteraction from '@antv/g-plugin-dom-interaction';
 import * as HTMLRenderer from '@antv/g-plugin-html-renderer';