+++ packages/g-pattern/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @antv/g-pattern
 
+## 1.2.7
+
+### Patch Changes
+
+-   70aa0b32: Retrieve runtime from context instead of global.
+-   789bd4c9: Split a path containing multiple segments into subpaths.
+-   Updated dependencies [70aa0b32]
+-   Updated dependencies [789bd4c9]
+    -   @antv/g-lite@1.2.7
+
 ## 1.2.6
 
 ### Patch Changes