+++ packages/g-plugin-a11y/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-plugin-a11y
 
+## 0.6.7
+
+### Patch Changes
+
+-   Updated dependencies [70aa0b32]
+-   Updated dependencies [789bd4c9]
+    -   @antv/g-lite@1.2.7
+
 ## 0.6.6
 
 ### Patch Changes