+++ packages/g-plugin-annotation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-plugin-annotation
 
+## 0.4.7
+
+### Patch Changes
+
+-   Updated dependencies [70aa0b32]
+-   Updated dependencies [789bd4c9]
+    -   @antv/g-lite@1.2.7
+
 ## 0.4.6
 
 ### Patch Changes