+++ packages/g-plugin-canvas-path-generator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @antv/g-plugin-canvas-path-generator
 
+## 1.3.7
+
+### Patch Changes
+
+-   70aa0b32: Retrieve runtime from context instead of global.
+-   789bd4c9: Split a path containing multiple segments into subpaths.
+-   Updated dependencies [70aa0b32]
+-   Updated dependencies [789bd4c9]
+    -   @antv/g-lite@1.2.7
+    -   @antv/g-math@2.0.2
+
 ## 1.3.6
 
 ### Patch Changes