+++ packages/g-plugin-webgpu-device/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @antv/g-plugin-webgpu-device
 
+## 1.9.7
+
+### Patch Changes
+
+-   70aa0b32: Retrieve runtime from context instead of global.
+-   789bd4c9: Split a path containing multiple segments into subpaths.
+-   Updated dependencies [70aa0b32]
+-   Updated dependencies [789bd4c9]
+    -   @antv/g-plugin-device-renderer@1.9.7
+    -   @antv/g-lite@1.2.7
+
 ## 1.9.6
 
 ### Patch Changes