+++ packages/g-web-components/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @antv/g-web-components
 
+## 1.9.7
+
+### Patch Changes
+
+-   Updated dependencies [70aa0b32]
+-   Updated dependencies [789bd4c9]
+    -   @antv/g-canvas@1.11.7
+    -   @antv/g-webgl@1.9.7
+    -   @antv/g-lite@1.2.7
+
 ## 1.9.6
 
 ### Patch Changes